Split stuff out of wayland_connector.cpp to fix Fedora 27 build (shaves a slice off issue #132)
