+++ src/server/frontend/wayland/CMakeLists.txt
@@ -3,8 +3,8 @@ set(
 
   core_generated_interfaces.h
   wayland_default_configuration.cpp
-  wayland_connector.cpp
-)
+  wayland_connector.cpp wayland_connector.h
+  wlshmbuffer.cpp       wlshmbuffer.h)
 
 add_library(
   mirfrontend-wayland OBJECT
